add service and rename
The cluster service names are to restrictive and long.
And they cannot stop the real service
that the virtual cluster-service
is meant to start and provide.
One service is added that depends on
the former cluster-service scripts
which are renamed.
They probably need to be renamed again...

remove wrong file