fix poweroff issue
some nodes do not come up when they are shutdown properly by
the default cmd "openrc-shutdown -p now".
These nodes come up when they get power on the line after loosing power during operation befor.
So the solution is to not power them off and only halt them.
Then they are taken of the powerline and will be able to come up if the power is
reapplied.